ppc/ppc405: QOM'ify CPC
Introduce a QOM property "cpu" to initialize the DCR handlers. This is
a pattern that we will reuse for the all other 405 devices needing it.

Now that all clock settings are handled at the CPC level, change the
SoC "sys-clk" property to be an alias on the same property in the CPC
model.

Reviewed-by: Daniel Henrique Barboza <danielhb413@gmail.com>
Signed-off-by: Cédric Le Goater <clg@kaod.org>
Message-Id: <20220803132844.2370514-8-clg@kaod.org>
[danielhb: removed redundant typedef declaration]
Signed-off-by: Daniel Henrique Barboza <danielhb413@gmail.com>